+++ firmware/common/max2837.h
@@ -1,8 +1,32 @@
+/*
+ * Copyright 2012 Will Code? (TODO: Proper attribution)
+ * Copyright 2014 Jared Boone <jared@sharebrained.com>
+ *
+ * This file is part of HackRF.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; see the file COPYING.  If not, write to
+ * the Free Software Foundation, Inc., 51 Franklin Street,
+ * Boston, MA 02110-1301, USA.
+ */
+
 #ifndef __MAX2837_H
 #define __MAX2837_H
 
 #include <stdint.h>
 #include <stdbool.h>
+
+#include "max2837_drv.h"
 
 /* TODO - make this a private header for max2837.c only, make new max2837.h */
 
@@ -29,20 +53,10 @@ extern uint16_t max2837_reg_read(uint8_t r);
  * clean. */
 extern void max2837_reg_write(uint8_t r, uint16_t v);
 
-/* Read all registers from chip and copy to memory. Mark all clean. */
-extern void max2837_regs_read(void);
-
 /* Write all dirty registers via SPI from memory. Mark all clean. Some
  * operations require registers to be written in a certain order. Use
  * provided routines for those operations. */
 extern void max2837_regs_commit(void);
-
-typedef enum {
-	MAX2837_MODE_SHUTDOWN,
-	MAX2837_MODE_STANDBY,
-	MAX2837_MODE_TX,
-	MAX2837_MODE_RX
-} max2837_mode_t;
 
 void max2837_mode_shutdown(void);
 void max2837_mode_standby(void);
