+++ firmware/hackrf-common.cmake
@@ -135,6 +135,7 @@ macro(DeclareTargets)
 		${PATH_HACKRF_FIRMWARE_COMMON}/rf_path.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/si5351c.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/max2837.c
+		${PATH_HACKRF_FIRMWARE_COMMON}/max2837_drv.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/max5864.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/rffc5071.c
 		m0_bin.s
